requirements.txt
django==4.2.11
django-environ==0.11.2
django-filter==24.2
django-import-export==3.3.8
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-standardized-errors==0.13.0
mysqlclient==2.2.4
pandas==2.0.3